invoke toString for objects. handles special case of arrays; arrays of objects are supported (the objects toString is invoked).
